+++ PowerPointLabs/PowerPointLabs/ColorThemes/ThemeManager.cs
